Rename file to avoid confusing name collision.
